cidade em polos e escola para cadastro inicial

+++ envio_cadastro_polo.php
@@ -8,6 +8,7 @@ $enderecoFuncionamento = $_POST["enderecoFuncionamento"];
 $localFuncionamento = $_POST["localFuncionamento"];
 $horaFuncionamento = $_POST["horaFuncionamento"];
 $diaFuncionamento = $_POST["diaFuncionamento"];
+$idCidade = $_POST["idCidade"];
 
 
 
@@ -18,11 +19,13 @@ if(mysqli_num_rows($sql) > 0){
 	echo "<script>alert('Polo já existente! Cadastre um Polo novo');window.location='cadastrar_polo.php'</script>";
 exit();
 } else {
- !$con->query("INSERT INTO polo (nomePolo, dtCriacao, enderecoFuncionamento, localFuncionamento, horaFuncionamento, diaFuncionamento, status, teste, teste2 ) VALUES ('$nomePolo', '$dtCriacao', '$enderecoFuncionamento',
- '$localFuncionamento', '$horaFuncionamento', '$diaFuncionamento', 1  )");
- echo "<script>alert('Cadastro realizado com sucesso!');window.location='cadastrar_polos.php'</script>";
+ $confirma = $con->query("INSERT INTO polo (nomePolo, dtCriacao, enderecoFuncionamento, localFuncionamento, horaFuncionamento, diaFuncionamento, status, idCidade ) VALUES ('$nomePolo', '$dtCriacao', '$enderecoFuncionamento',
+ '$localFuncionamento', '$horaFuncionamento', '$diaFuncionamento', 1, '$idCidade'  )");
+ 
+ if($confirma === true){
+	echo "<script>alert('Cadastro realizado com sucesso!');window.location='cadastrar_polos.php'</script>";
+ }else{
+	echo "Erro para inserir: " . $con->error;
+ }
 }
 $con->close();
-
-
-?>
